requirements.txt
pandas==2.0.1
wntr==1.0.0
xlwt==1.3.0
click==8.1.3
click-plugins==1.1.1
PyYAML==6.0
click
rich-click
